Move hook tests up one directory
