+++ __tests__/setupHooks/emptyHookFile.js
@@ -1,8 +1,7 @@
 const fs = require("fs");
 const { resolve } = require("path");
 let debug = require("debug");
-const { loadSetupHooks } = require("../../../hooksLoadout");
-const { removeConfig } = require("../../../tests/fixtures/configWriter");
+const { loadSetupHooks } = require("../../hooksLoadout");
 
 jest.mock("debug", () => {
     let originalModule = jest.requireActual("debug");
@@ -10,12 +9,15 @@ jest.mock("debug", () => {
     return originalModule;
 });
 
-const filePath = resolve(__dirname, "../../../setupHooks.js");
+const filePath = resolve(__dirname, "../../setupHooks.js");
 debug.enable("jest-mysql:hooksLoadout");
 
-it("Should not to load hooks if provided with empty object", async () => {
+beforeAll(async () => {
     await writeEmptyFileExport(filePath);
+    jest.resetModules();
+});
 
+it("Should not to load hooks if provided with empty object", async () => {
     await loadSetupHooks();
 
     expect(debug.log.mock.calls[1][0]).toMatch(/No setup hooks provided/);
@@ -25,11 +27,3 @@ async function writeEmptyFileExport(filePath) {
     const file = "module.exports = {};";
     await fs.promises.writeFile(filePath, file);
 }
-
-async function removeEmptyFileExport(filePath) {
-    await fs.promises.unlink(filePath);
-}
-
-afterAll(async () => {
-    await removeConfig(filePath);
-});
